Lowercase address when checking assoc state
