+++ bindings_node/src/inbox_id.rs
@@ -1,8 +1,11 @@
 use crate::ErrorWrapper;
 use napi::bindgen_prelude::Result;
+use napi::bindgen_prelude::Uint8Array;
 use napi_derive::napi;
+use std::sync::Arc;
 use xmtp_api_grpc::grpc_api_helper::Client as TonicApiClient;
 use xmtp_id::associations::generate_inbox_id as xmtp_id_generate_inbox_id;
+use xmtp_id::associations::MemberIdentifier;
 use xmtp_mls::api::ApiClientWrapper;
 use xmtp_mls::retry::Retry;
 
@@ -36,4 +39,54 @@ pub fn generate_inbox_id(account_address: String) -> Result<String> {
   // create_client function above, which also has a hard-coded nonce of 1
   let result = xmtp_id_generate_inbox_id(&account_address, &1).map_err(ErrorWrapper::from)?;
   Ok(result)
+}
+
+#[napi]
+pub async fn is_installation_authorized(
+  host: String,
+  inbox_id: String,
+  installation_id: Uint8Array,
+) -> Result<bool> {
+  is_member_of_association_state(
+    &host,
+    &inbox_id,
+    &MemberIdentifier::Installation(installation_id.to_vec()),
+  )
+  .await
+}
+
+#[napi]
+pub async fn is_address_authorized(
+  host: String,
+  inbox_id: String,
+  address: String,
+) -> Result<bool> {
+  is_member_of_association_state(
+    &host,
+    &inbox_id,
+    &MemberIdentifier::Address(address.to_lowercase()),
+  )
+  .await
+}
+
+async fn is_member_of_association_state(
+  host: &str,
+  inbox_id: &str,
+  identifier: &MemberIdentifier,
+) -> Result<bool> {
+  let api_client = TonicApiClient::create(host, true)
+    .await
+    .map_err(ErrorWrapper::from)?;
+  let api_client = ApiClientWrapper::new(Arc::new(api_client), Retry::default());
+
+  let is_member = xmtp_mls::identity_updates::is_member_of_association_state(
+    &api_client,
+    inbox_id,
+    identifier,
+    None,
+  )
+  .await
+  .map_err(ErrorWrapper::from)?;
+
+  Ok(is_member)
 }
